requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
faiss-cpu==1.8.0
filelock==3.13.4
fsspec==2024.3.1
huggingface-hub==0.22.2
idna==3.7
Jinja2==3.1.3
joblib==1.4.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
nltk==3.8.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
packaging==24.0
pandas==2.0.3
pillow==10.3.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==2.7.0
six==1.16.0
sympy==1.12
threadpoolctl==3.4.0
tokenizers==0.19.1
torch==2.2.2
tqdm==4.66.2
transformers==4.40.1
triton==2.2.0
typing-extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
websockets==12.0
